feat: continue to register with social
