+++ packages/core/src/lib/sign-in.ts
@@ -1,8 +1,9 @@
 import { PasscodeType, UserLogType } from '@logto/schemas';
 import { Context } from 'koa';
-import { InteractionResults, Provider } from 'oidc-provider';
+import { Provider } from 'oidc-provider';
 
 import { getSocialConnectorInstanceById } from '@/connectors';
+import { SocialUserInfo } from '@/connectors/types';
 import RequestError from '@/errors/RequestError';
 import { WithUserLogContext } from '@/middleware/koa-user-log';
 import {
@@ -18,7 +19,7 @@ import assertThat from '@/utils/assert-that';
 import { emailRegEx, phoneRegEx } from '@/utils/regex';
 
 import { createPasscode, sendPasscode, verifyPasscode } from './passcode';
-import { getUserInfoByConnectorCode, getUserInfoFromInteractionResult } from './social';
+import { SocialUserInfoSession } from './social';
 import { findUserByUsernameAndPassword } from './user';
 
 const assignSignInResult = async (ctx: Context, provider: Provider, userId: string) => {
@@ -119,26 +120,38 @@ export const assignRedirectUrlForSocial = async (
   ctx.body = { redirectTo };
 };
 
+const saveUserInfoToSession = async (
+  ctx: Context,
+  provider: Provider,
+  socialUserInfo: SocialUserInfoSession
+) => {
+  const redirectTo = await provider.interactionResult(
+    ctx.req,
+    ctx.res,
+    {
+      socialUserInfo,
+    },
+    { mergeWithLastSubmission: true }
+  );
+  ctx.body = { redirectTo };
+};
+
 export const signInWithSocial = async (
   ctx: WithUserLogContext<Context>,
   provider: Provider,
-  { connectorId, code, result }: { connectorId: string; code: string; result?: InteractionResults }
+  connectorId: string,
+  userInfo: SocialUserInfo
 ) => {
   ctx.userLog.connectorId = connectorId;
   ctx.userLog.type = UserLogType.SignInSocial;
 
-  const userInfo =
-    code === 'session'
-      ? await getUserInfoFromInteractionResult(connectorId, result)
-      : await getUserInfoByConnectorCode(connectorId, code);
-
-  assertThat(
-    await hasUserWithIdentity(connectorId, userInfo.id),
-    new RequestError({
+  if (!(await hasUserWithIdentity(connectorId, userInfo.id))) {
+    await saveUserInfoToSession(ctx, provider, { connectorId, userInfo });
+    throw new RequestError({
       code: 'user.identity_not_exists',
       status: 422,
-    })
-  );
+    });
+  }
 
   const { id, identities } = await findUserByIdentity(connectorId, userInfo.id);
   // Update social connector's user info
