+++ packages/core/src/lib/social.ts
@@ -27,12 +27,12 @@ const getConnector = async (connectorId: string) => {
   }
 };
 
-export const getUserInfoByConnectorCode = async (
+export const getUserInfoByAuthCode = async (
   connectorId: string,
-  code: string
+  authCode: string
 ): Promise<SocialUserInfo> => {
   const connector = await getConnector(connectorId);
-  const accessToken = await connector.getAccessToken(code);
+  const accessToken = await connector.getAccessToken(authCode);
 
   return connector.getUserInfo(accessToken);
 };
@@ -41,15 +41,20 @@ export const getUserInfoFromInteractionResult = async (
   connectorId: string,
   interactionResult?: InteractionResults
 ): Promise<SocialUserInfo> => {
-  const result = z
+  const parse = z
     .object({
       socialUserInfo: z.object({
         connectorId: z.string(),
         userInfo: socialUserInfoGuard,
       }),
     })
-    .parse(interactionResult);
+    .safeParse(interactionResult);
 
+  if (!parse.success) {
+    throw new RequestError('session.connector_session_not_found');
+  }
+
+  const result = parse.data;
   assertThat(result.socialUserInfo.connectorId === connectorId, 'session.connector_id_mismatch');
 
   return result.socialUserInfo.userInfo;
