+++ packages/phrases/src/locales/en.ts
@@ -65,6 +65,7 @@ const errors = {
     invalid_connector_id: 'Unable to find available connector with id {{connectorId}}.',
     insufficient_info: 'Insufficent sign-in info.',
     connector_id_mismatch: 'The connectorId is mismatched with session record.',
+    connector_session_not_found: 'Connector session not found. Please go back and sign in again.',
   },
   connector: {
     general: 'An unexpected error occured in connector.',
